Updated to use private as main source, if available

+++ bash/install.sh
@@ -23,7 +23,12 @@ install() {
   ln -s ~/.dotfiles/bash/symlink.bashrc ~/.bashrc
 
   rm ~/.bash_profile
-  ln -s ~/.dotfiles/bash/symlink.bash_profile ~/.bash_profile
+
+  if [ -f ~/.dotfiles/dotfiles_private/.bash_profile ]; then
+    ln -s ~/.dotfiles/dotfiles_private/.bash_profile ~/.bash_profile
+  else
+    ln -s ~/.dotfiles/bash/symlink.bash_profile ~/.bash_profile
+  fi
 }
 
 update() {
